move DesktopNavBar to a different folder, seperated components to Nav, implemented DesktopNavBar wity Portal component
