+++ src/components/UI/Nav/Nav.module.css
@@ -54,7 +54,6 @@
 	left: 0;
 	background-color: var(--color-navbar);
 	box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-	border-radius: 4px;
 	padding: 0px;
 	z-index: 1;
 	margin: 0;
