+++ src/components/UI/index.js
@@ -1,11 +1,8 @@
+import { Nav, NavItem } from "./Nav/Nav";
 import Button from "./Button/Button";
 import Input from "./Input/Input";
 import Label from "./Label/Label";
 import TextArea from "./TextArea/TextArea";
+import Portal from "./Portal/Portal";
 
-export { 
-	Button, 
-	Input, 
-	Label, 
-	TextArea 
-};
+export { Button, Input, Label, TextArea, Nav, NavItem, Portal };
